Fix: Show Selected Feed's Details (Compiler Update)
When selecting a feed in the FeedGrid, the FeedDetails were not updated.
This happened because the onSelect event was not handled due to an MXML
compiler bug. The compiler resolved FeedGrid#onSelect property assigned in
FeedDetail.mxml to the private method instead of the public [Event] declared
in FeedGrid's super class TablePanel. This is a bug, since methods can never be
assigned / overwritten, and private members should also be skipped.
This bug is now fixed, so all we have to do is bump dependencies to the new
Jangaroo compiler version.