Adapt hello-world-localized Example to Changed Jangaroo API
In pom.xml, <additionalLocales> has to be used instead of <locales>
(because you need not repeat the default locale "en").
Customized files before-ext-load.js and index.html is no longer needed.
Instead of localization-config.js, you now have to specify a file
resources/config-init.js with the supported locales in a JSON object.
This could even be generated, because it usually contains exactly the
same list already given in pom.xml <additionalLocales>.